Revert "[mlir][tosa] Add support for dense_resource in tosa-narrow-* passes (…"
This reverts commit 4053765e4290266ec640c13dcd2920c3bab96f72.
