+++ mlir/test/Dialect/Tosa/tosa-narrow-f64-to-f32.mlir
@@ -178,23 +178,3 @@ func.func @test_f64_add_diagnostic(%arg0: tensor<13x21x1xf64>, %arg1: tensor<13x
   return %0 : tensor<13x21x3xf64>
 }
 }
-
-// -----
-
-// CHECK-LABEL: test_dense_ressource_f64
-func.func @test_dense_ressource_f64() -> tensor<1x2xf64> {
-  // COMMON: %[[CONST:.*]] = "tosa.const"() <{values = dense_resource<resource> : tensor<1x2xf32>}> : () -> tensor<1x2xf32>
-  %0 = "tosa.const"() <{values = dense_resource<resource> : tensor<1x2xf64>}> : () -> tensor<1x2xf64>
-  // DEFAULT: %[[OUT_CAST:.*]] = tosa.cast %[[CONST]] : (tensor<1x2xf32>) -> tensor<1x2xf64>
-  // DEFAULT: return %[[OUT_CAST]] : tensor<1x2xf64>
-  // FUNCBOUND: return %[[CONST]] : tensor<1x2xf32>
-  return %0 : tensor<1x2xf64>
-}
-{-#
-  dialect_resources: {
-    builtin: {
-      // COMMON: resource: "0x040000000000803F000080BF"
-      resource: "0x04000000000000000000F03F000000000000F0BF"
-    }
-  }
-#-}
